+++ source/Core/Model/BaseModel.php
@@ -763,9 +763,27 @@ class BaseModel extends \OxidEsales\Eshop\Core\Base
      */
     public function getFieldData($fieldName)
     {
+        return ($field = $this->getFieldIfAvailable($fieldName)) ? $field->value : null;
+    }
+
+    /**
+     * Gets raw field data
+     *
+     * @param string $fieldName name (eg. 'oxtitle') of a data field to get
+     *
+     * @return mixed value of a data field
+     *
+     */
+    public function getRawFieldData($fieldName)
+    {
+        return ($field = $this->getFieldIfAvailable($fieldName)) ? $field->rawValue : null;
+    }
+
+    private function getFieldIfAvailable($fieldName): ?Field
+    {
         $longFieldName = $this->getFieldLongName($fieldName);
 
-        return ($this->$longFieldName instanceof Field) ? $this->$longFieldName->value : null;
+        return ($this->$longFieldName instanceof Field) ? $this->$longFieldName : null;
     }
 
     /**